Fix error when saving last submodule section.
Currently we unconditionally try to write a submodule's branch configuration,
even when no branch is set. This can cause an error in libgit2 when we try to
save a submodule's configuration when this submodule is the last section in the
corresponding file. Fix this error by first checking if the submodule has a
branch.