Fixed checkout to correctly handle executable bit changes
The current implementation of checkout wasn't taking into account
file mode changes between blob and executable blob when looking
at changes in the workdir. This means that if a file was executable
in the index but not in the workdir, even a force checkout would not
update the file in the workdir.