remote: refuse to load multi-url remotes
A remote can have more than a single entry in url or pushurl. Git uses
this to perform multiple fetches or pushes, which can be useful for
redundancy, but we currently cannot handle this.

What we would do is fetch from or push to the last of the urls (as we
consider them single-value fields), which would at best be confusing,
but would also break the user's expectations.

in order to avoid this, check whether the remote we're trying to lookup
has mutliple urls and refuse to load it.