Lock around encrypted I/O
OpenSSL and often whatever libssh2 is using require their own set-up for
concurrent operations to be safe. This means that by default, using both
of these libraries in a threaded context is unsafe.

Lock by default, and allow the user to tell us that they've set up
threading for the underlying libraries.