buffer: Fix growing non-allocated buffers to modulo 8 sizes
git_buf_grow() ensures the buffer is terminated with a NUL byte, but
dares to truncate the buffer if it is too short.  So, allocation has
to always be 1 more than the required content size.

This is generally not a problem because for already allocated buffers
the grow logic makes it impossible for the content size to be equal to
the allocated size -- unless the target size is the same as the
allocated size, but this has a special early-out handling.

However, for non-allocated buffers, the allocation is exactly the
requested target size if the target size is a multiple of the alignment
(8), in which case truncation will occur if the content size is the
same as the requested target size.

Fix this by allocating one more byte than actually requested when the
buffer isn't already allocated.