errors: remove the error "class"
The error "class" is a poorly defined construct and not something
that is particularly reliable for callers.  Remove it.

The class was used internally to signal an operating system error,
and to provide the operating system's error message information.
Instead, provide `giterr_set_os` for this functionality.  The class
was also used to identify OOM situations.  Instead, provide
`giterr_is_oom` to identify these.