Patched maybe_append_head() so it doesn't return an error if there's no head.
The documentation for maybe_append_head() says it’ll append to the head
if the head is pointing at the thing just modified. In the case where
there is NO head this function was returning an error, which got
bubbled up to Objective-Git and was causing what appeared to be
failures when there weren’t.

It is vacuously true that if there is no head reference then it does
NOT point to anything that has been updated and thus
maybe_append_head() has no work to do and shouldn’t throw an error
(since it’s called “maybe” not “you_must_append_head_or_die”).