CMakeLists: Equalize comparisons to CMAKE_SIZEOF_VOID_P
Compare CMAKE_SIZEOF_VOID_P as a number, not as a string. Also,
consistently use 32-bit as the fallback to work around combinations of
CMake / compiler versions where CMAKE_SIZEOF_VOID_P is unset. This should
be uncritical as there are no relevant architectures (yet) that are
neither 32- nor 64-bit, and even if so, they are still likely to be able
to run 32-bit code, too.