Fix concurrently calling git_sysdir_get
If this function was concurrently called, it was possible for all threads
entering to call `git__on_shutdown` which could possibly trigger the assert that
there aren't too many shutdown handlers (as each thread entering would try to
register one).

This adds an atomic variable to count the number of threads entering, and only
the first will register a shutdown handler. Additionally, the shutdown handler
no longer flags itself as not having been set, as it is indeed still set.